pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chessbrain-content"
version = "0.1.0"
description = "Automated daily social content for Chess Brain (Lichess AI coach Discord bot)"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "tenacity>=8.2",
    "apscheduler>=3.10",
    "python-chess>=1.999",
    "cairosvg>=2.7",
    "fal-client>=0.5",
    "groq>=0.11",
    "openai>=1.40",
    "numpy>=1.26",
    "moviepy>=2.0",
    "zstandard>=0.22",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.5"]

[project.scripts]
chessbrain = "chessbrain.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 110
target-version = "py311"
